migrated select test to test_schema.py and rename apply to select
